cache Modules that compiled w/ or w/o error
Before, if there were any errors, we discarded all modules. But this
needlessly caused us to have to keep re-compiling a root node until in
a dependency tree until all leaf notes were fixed, which is no good.

Further, if cache even if there are errors. It makes no sense to try
again until the source or the source of some dependency changes.